v0a1
1. Отображается график и можно редактировать его
2. Переделона сетка
3. Много других исправлений...

+++ WpfApplication4/Oscil.cs
@@ -14,7 +14,7 @@ namespace WpfApplication4
         public static DateTime StampDateStart;
         public static DateTime StampDateTrigger;
         public static DateTime StampDateEnd;
-        public static uint SampleRate;
+        public static double SampleRate;
 
         public static ushort ChannelCount;
         public static uint NumCount;
